Changed the protocol so that PROGRAM is no longer part of a standard request.
sendbackup is no longer split into two different programs.  Just one
sendbackup program is built.  It will lookup a program table for the
PROGRAM name specified for the filesystem and obtain the other
relevant data, such as regexp's, backup and restore programs and start
and end functions from this table.  This eases integrating different
backup methods into sendbackup.  The integration of them into sendsize
is left as an exercise to the developer :-)

sendsize now supports concurrent estimation of backups, estimating
concurrently at most as many filesystems as specified in the maxdumps
field of the dumptype.  Locking of the standard output of multiple
subprocesses is taken care of in sendsize.c and calcsize.c