Removed the dist-hook from the main Makefile.am (it failed to cp {docs,patches}/* from $(srcdir)/).  Adding $(srcdir), it would try to copy the CVS directory and would fail too.
Created Makefiles for docs and patches just for creating the
distribution correctly.  Unfortunately, new docs and patches files
will have to be added manually to Makefile.am.

Removed some duplicates from common-src/Makefile.am

Autogen is no longer necessary.  Its functionality is already provided
in autoconf's autoreconf script.  It requires a patch in order to
support automake files, though.  This patch was included in
patches/autoreconf.patch (but it is not supposed to be added to the
distribution)