Fixed amandates so that /etc/amandates is created with 0644 access, instead of 0666.  (suggested by Stefan Monnier <monnier@TEQUILA.SYSTEMSZ.CS.YALE.EDU>)
In configure.in, default index directory and gnutar list directory
changed from amanda-index and amanda-gnutar-lists to amanda/index and
amanda/gnutar-lists.  (also suggested by Stefan Monnier)

Created variable DB_EXT in configure.in that will be a list of the
extensions used by a database package.  For dbm and ndbm, it is `.pag
.dir'; for db it is `.db' and for gdbm it is `""'.  It can be used as
the list for a for command in a shell script.

Fixed amrmtape.sh so that it will look for the correct db filenames.
(pointed out by Stefan Monnier)

Fixed configure.in so that, if any essential server dependency is
missing (locking support, for instance), the behavior is exactly the
same as if --without-server had been issued.