Change the sort order for schedq so that it is now by decreasing size within priority levels.
Now small low-priority dumps will be picked on first.  This should have
the effect of filling the tape with a few big dumps rather than part filling
the tape with many small dumps.

Naturally, once the plan includes as many large dumps as will fit many of
the little dumps will be reinstated to fill the remaining hole.