The DOC on index say that only the leaf are on the index file, but tar also include the directories. It's is a good idea to include the directories since it allow amrecover to recover an empty directory which will not be listed otherwise. I have changed sendbackup-dump.c to include the directories in index.
The current directory look like an empty line in amrecover, I changed it
to a dot ".", you can use the command "add ." to add the current
directory, this is usefull for adding the entire dump.

The filter that read index file remove the trailing /, I changed it
to keep it, amrecover will write the /, that will differentiate
files from directories.

Amindexd uncompressed, grep and sort index file on each request.
I changed it to uncompress and sort only on the first request,
the sorted files are keep on disk and are erased at program quit.

amindexd now check that the status of the dump is "OK" from
the output of 'amadmin find' command, it will not consider failed dump.

struct DIR_ITEM in list_dir.c use 'char *path' instead
of 'char path[1024]', that save a lot of memory.

I corrected the OISD command of amindexd so that it check this
is really a directory by adding a / at the end of the path. The cd
command of amrecover will only cd to a directory, not a file or a part
of a file. (if you have the file foobar, it was possible to "cd foo").

I added the ORLD (Opaque recursive list directory) to amindexd, it
is similar to OLSD but is recursive.

I changed the get_line() function of amrecover to read a buffer instead
of just one char, this can be 10 times faster when the result
is many lines (output of ORLD command).

When you add a directory in amrecover, it will execute a ORLD command to
find all tape needed to restore the directory and its content. It
will not only restore from the last tape.

Restoring from multiple tape is done in increasing date order.

Before amidxtaped execv amrestore, it change the stderr to it's db_file.
amrestore will write it's error message to the debug file of amidxtaped.

by Jean-Louis Martineau <martinea@iro.umontreal.ca>