CVS: ---------------------------------------------------------------------- automatically CVS: CVS: Committing in .  CVS: CVS: Modified Files: amanda-2/config/acconfig.h amanda-2/server-src/conffile.c CVS: amanda-2/server-src/diskfile.c amanda-2/server-src/dumper.c CVS: amanda-2/server-src/taper.c CVS: ---------------------------------------------------------------------- Instead of having each .c or .h file that uses the GNU C __attribute__ do an #ifdef, do an #ifdef in amanda.h and if the compiler is not GNU C, then #define __attribute__ away.
In amanda.h, if HAVE_SNPRINTF is defined, then define ap_snprintf to
snprintf and ap_vsnprintf to vsnprintf.  We should use the functions
ap_snprintf and ap_vsnprintf throughout our code.  Declare ap_snprintf
and ap_vsnprintf if snprintf does not exist.  Declare snprintf and
vsnprintf if snprintf exists but is not declared.