Make sure internal errors, such as the mail command failing, get reported.
In the case of amcheck, they go to stderr even if -m is set since there
didn't seem to be anyplace better to do so.  If this is running from
a cron job it might get back to a real user.

For reporter, the errors go to .../amdump along with everything else from
the amdump script, and also to .../log, except since it's reporter that's
having the trouble, they probably will not make it into the mail report.