A change was made between 2.4.1p1 and 2.4.2 to the way amdump detects who is running it.  The idea was to handle systems that do not have the "whoami" program.
The change used the $LOGNAME environment variable if defined, then fell
back to "whoami".  But since LOGNAME is not updated by "su" unless you use
the '-' option, and in general is not as reliable as actually checking
the real uid, the following patch reverses the check so we try "whoami"
first (but allow for it to be absent), then "id" (also allowing for it
to be absent), then $LOGNAME, then gives up.

With thanks to Jon LaBadie <jon@jgcomp.com> for his insight on this.

This is a temporary fix.  The real fix will involve changing the various
server programs to each check for themself.