Move the device name parsing routines from the RAIT module to tapeio and make them generic.  This allows one module to open another, e.g. rait:{null:,z1,z2,z3}.
Maintain both the original open flags and mask and use
them on the real tapeio and driver open.

Make sure amdd and ammt compile "stand-alone".  To do
so, use:

  cc -DNO_AMANDA amdd.c output-rait.c output-tape.c -o rait-dd
  cc -DNO_AMANDA ammt.c output-rait.c output-tape.c -o rait-mt

The output-tape.c routines are now included because
they know how to do the real tape ioctl calls on all the
OS types.

Abort as soon as one of the amdd open calls fails instead
of waiting for read/write to report a bad file descriptor.