change protocol_sendreq parameters to accept an optional callback function that will get passed through security_connect (parameters also changed) that can be used to find server configuration information.  This is done because the security functions are in common-src, but the configuration can only be queried from server-src modules (makes sense).
write a generic generic_get_security_conf to pass to security layers
that just interfaces simpley with the configuration file.  (This
could easily be used with other security layers if desired).

change the krb5 security module to not use hardcoded values for
principal and keytab, but instead make them configurable based on
values in amanda.conf.  The config values are accessible always, not just
when krb5 is built in (but are obviously useless).  This allows a site to
use the same config for both types of sites.

change the amanda authorization routines (common-src/krb5-security.c) to
fallback to kerberized-rsh like behavior if no .k5amandahosts file exists.
Also change krb5-security.c to use .k5login if USE_AMANDAHOSTS is not
defined.

A future enhancement would be the possibility to move principal/keytab
information to dumptypes so that different users can be used against
different hosts.  This would require being able to pass a more complex
configuration query callback down through the various layers.