match.c: match_word(): remove unused assignment
lenword is assigned before computing mword, and is assigned again after mword is
computed - to the exact value it was set to before mword is computed.

And even if the assignment was meant to be strlen(mword), the length would be
the same anyway as the substitution simply operates a character substitution.