match.c: move glob/disk backslash substitution out of match_word()
It really belongs to match_disk(). The original test in match_word():

if (separator == '/' && lenword > 2 && word[0] == '\\' && word[1] == '\\'
     && !strchr(word, '/')) {

really only checks whether a disk is a Windows share. So, make match_disk()
handle this check (which removes the need to check whether the separator is '/'
- it is, since we're checking a disk here) and also allocate (and free) the
"corrected" glob and disk if need be.

One advantage is that the mword variable in match_word() isn't needed anymore,
which is all the more useful that it held a full copy of the word argument - for
no purpose. Remove it entirely.