Ensure all tests return gbooleans
Most, if not all, tests returned ints, and quite a few of them returned TRUE or
FALSE. While the effect is basically the same as a gboolean, make all tests
return a gboolean for coherency, since this is what callinfork() expects anyway.

Also enforce that callinfork() return that type, which, in turn, allows the
"success" local variable in testutil_run_tests() to also be converted.