testutils.c: report time of execution for each test
The time is reported in seconds, with 3 digits after the decimal separator,
effectively reporting milliseconds.

In case a test fails, it is reported which iteration failed, and the total time
spent so far.

In case the number of occurrences for each test is strictly greater than 1, the
avg/min/max run is reported along with the total time spent.