match.c: factorize regex compile
regcomp() is used in many places in the file. Create a do_validate_regex()
function to do the job instead, and use it where appropriate.

This function returns a gboolean (TRUE if valid, FALSE otherwise) and fills an
error buffer (regex_errbuf, basically a char[]) with the error message if the
regex is invalid, so that the caller can use the error message.