match.c: better handling of "separator only" globs in match_word()
Extract "separator only" glob detection from match_word() into a separate
function. Also, the space for a full blown regex was allocated even if the glob
given as an argument was "separator only". Change that by allocating separately
in the two branches.

Also, get rid of the mglob variable: it is not used anymore. This avoids a
full-blown copy of the glob argument.