match.c: rework regex generation from globs
The process to generate a regex out of a glob is pretty much always the same:
append a beginning for the regex, then process the core with some special cases,
then append an end.

Generify all this by creating a amglob_to_regex() function which handles all
cases and use it in the appropriate functions: glob_to_regex(), tar_to_regex()
and match_word().