An implementation for cached buffers (xbufs)
The current xfer architecture has two mechanisms, XFER_MECH_PUSH_BUFFER and
XFER_MECH_PULL_BUFFER, by which XferElements pass themselves buffers of data.
The problem is that the upstream element (PUSH) will, in most cases, allocate buffers
dynamically before handling them down to the downstream element, where they will
be freed. If lots of buffers are exchanged this way, this means as many
*alloc()s, then *free()s.

Proposed solution: implement a cache of dynamically allocated buffers, called
xbufs, with sizes of individual buffers being powers of 2 (from 4k to 64k right
now), and the associated mechanisms and implementations XFER_MECH_PUSH_XBUF and
XFER_MECH_PULL_XBUF. Progressively modify each XferElement and XferElementGlue
to implement this new mechanism, and eventually phase out _BUFFER mechanisms
completely.

Implementation: each cache maintains a pool of xbufs in the shape of a crude,
singly linked list of xbuf structures. When a cache is initialized, a number of
xbufs are preallocated and added to the pool. More xbufs will be created if
needed. Each allocated xbuf has a pointer back to the cache it was issued from.
Cache insertion and deletion is protected by a mutex.

The struct xbuf is as follows:

struct xbuf {
    gpointer data;
    size_t size;
    struct xbuf *next;
    struct xbuf_cache *cache;
};

Note that the "size" field will be set to the size request by xbuf_cache_get()
(see below), and NOT to the size of xbufs in this cache. This means that the
worst-case space wastage for a requested xbuf is (sz/2) - 1, where sz is the
size of buffers allocated by the cache it is issued from. However, some xbufs
will not be associated to any cache: this happens if the size requested is
greater than the maximum size handled by caches. Those are called "oversized"
xbufs, and their associated buffer is exactly the requested size.

Each cache is also limited in the total amount of memory used by xbufs in its
pool (1 MiB by default). This will not prevent more xbufs to be created should
the need arise, but an xbuf returned to a cache which has reached its maximum
allocation will be destroyed.

There are only two methods to interact with the cache, xbuf_cache_get() and
xbuf_cache_put():

- struct xbuf *xbuf_cache_get(gsize size):

Calculate the power of 2 immediately greater than or equal to "size", and grab
the corresponding cache. If no cache is found (oversized xbuf, see above),
return a straight xbuf. Otherwise, pop the first xbuf from the pool, if any, and
return it. If the pool is empty, allocate a new one and return it.

- void xbuf_cache_put(struct xbuf *xbuf):

See what cache the xbuf comes from. If none, destroy it. Otherwise, test if the
number of allocated entries exceeds the allowed maximum: if yes, destroy it.
Otherwise, push it back on the pool.

TODO:
- statistics
- dynamic cache resize in case of cache "thrashing"?
- "mass" get/put?