An implementation for cached buffers (xbufs)
The current xfer architecture has two mechanisms, XFER_MECH_PUSH_BUFFER and
XFER_MECH_PULL_BUFFER, by which XferElements pass themselves buffers of data.
The problem is that the upstream element (PUSH) will, in most cases, allocate
buffers dynamically before handling them down to the downstream element, where
they will be freed. If lots of buffers are exchanged this way, this means as
many *alloc()s, then *free()s.

Proposed solution: implement a cache of dynamically allocated buffers, called
xbufs, and the associated mechanisms and implementations XFER_MECH_PUSH_XBUF and
XFER_MECH_PULL_XBUF. Progressively modify each XferElement and XferElementGlue
to implement this new mechanism, and eventually phase out _BUFFER mechanisms
completely.

Implementation: each XferElement which needs one will create a cache upon
initialization; each cache maintains a pool of xbufs in the shape of a crude,
singly linked list of xbuf structures. When a cache is initialized, a number of
xbufs are optionally preallocated and added to the pool. More xbufs will be
created if needed. Cache insertion and deletion is protected by a mutex.

The struct xbuf is as follows:

struct xbuf {
    gpointer data;
    size_t bufsize;
    size_t size;
    struct xbuf *next;
};

The "size" field will be set to the size request by xbuf_cache_get()
(see below), and NOT to the size of xbufs in this cache.

The struct xbuf_cache is as follows:

struct xbuf_cache {
    GMutex *mutex;
    size_t size;
    struct xbuf *list;
    guint nr_entries;
};

An empty cache has its list member set to NULL.

The methods to create and destroy a cache are:

- struct xbuf_cache *xbuf_cache_create(size_t size, guint nr_entries)

Create a cache which will allocate "size" bytes to each xbuf it holds.
Initially create "nr_entries" xbufs and put them into the pool.

- void xbuf_cache_destroy(struct xbuf_cache *victim)

Destroy all xbufs in the pool, if any, and then the cache itself.

Thetwo methods to interact with the cache are:

- struct xbuf *xbuf_cache_get(struct xbuf_cache *cache, size_t size)

Grab an xbuf from the cache pool, set its size to "size". If the pool is empty,
create a new xbuf. Note that all xbufs will be created with the size of buffers
declared in the cache, not the requested size.

- void xbuf_cache_put(struct xbuf_cache *cache, struct xbuf *xbuf)

Store an xbuf into a cache. This MAY NOT be the original cache. However, the
xbuf's registered buffer size and the cache's block size MUST match.

TODO:
- statistics?
- "mass" get/put?
- memory limitation?