file.c: add am_full_read()
Gnulib's full_read(), despite having the same prototype as read(2), has awkward
semantics which make it very hard to work with: it requires to check errno if
the size returned is less than the requested size, since it may indicate an
error. Not only is this confusing, but it can lead to programming mistakes if it
is assumed that it works like read(2).

Create a simple wrapper, am_full_read(), which uses full_read() but behaves like
read(2) in case of error (ie, return -1).

NOTE: this only concerns C code. Perl is not affected, since it has its own
implementations of full_read/full_write using the POSIX package.