server-src/chunker.c: fix error checking
The code was reading:

	nr_read = full_read(fd, xxx, size);
	amclose(); /* twice */
	if (nr_read < size) {
		/* test errno */

The problem is that amclose calls close(2), which sets errno. So the errno
checking, which was meant to check the return of full_read(), was wrong.

Create a saved_errno variable and save the return code from full_read()
immediately after it has occurred, and check this value instead of errno.