server-src/amcheck.c: use am_full_read()
Two usages of full_read() within the file assumed that a return value strictly
greater than 0 meant success. But this is not the case. What's more, in one
usage, errno was checked after the full_read() loop too late to really matter if
full_read() failed before.

Change full_read() to am_full_read() in both cases. For the first occurrence,
test for -1 to exit early.