device-src/xfer-dest-taper-splitter.c: fix usage of full_read()
The bytes_read local variable was set to a size_t, and was used to hold the
value of a call to full_read(). However, it was tested for -1, which full_read()
can never return, or a value less than the requested bytes. Errors are not
checked for: am_full_read() is therefore not needed here.

Turn this variable into a gsize instead, which is unsigned and fills the
purpose, and remove the "less than 0" test.