ndmp-src/amndma_tape_simulator.c: use am_full_read()
The return of full_read() in ndmos_tape_read() was tested against values less
than 0, which full_read() can never return. Since the code makes it clear that
this was a test for a read error, use am_full_read() instead.

Also convert rc from being an int to an ssize_t, which better fits the purpose.