server-src/chunker.c: fix read error checking
In do_chunk(), the read error was checked for as such:

	nread = full_read(...);
	aclose(); /* twice */
	if (nread < xxx) {
		if (errno != 0) /* and so on */

The problem is that the errno check comes after two calls to amclose(), which
calls close(2), which sets errno. As a result, the test would check the wrong
value of errno.

Use read_fully() to capture errno right after the read operation and test the
captured value instead.