XferElement: add minimum, maximum and effective transfer sizes
Currently, the Xfer architectre does not know what the maximum permissible
transfer size will be along the transfer chain. This can lead to unnecessary
calls to memcpy() if, for instance, a downstream element sends 2*n sized buffers
to an upstream element which only accepts n sized buffers.

This patch brings the necessary modifications so that this maximum permissible
transfer size be known before the transfer effectively starts:

* an XferElement gains two new members, min_tsize and max_tsize, which define
  the minimum and maximum amount of data it can send; it also gains a
  get_blocksize() method to collect these values;

* it also gains a tsize member, and a set_transfersize() method which will set
  this value at transfer initialization time.

Implement stubs for XferElement. By default, the minimum and maximum transfer
sizes are 1 and GUINT_MAX.