Xfer: calculate, and enforce, the maximum individual transfer size
Create an xfer_set_transfersize() method, which goes through the list of
XferElements to calculate, and then set, the transfer size for this particular
Xfer. The transfer size is determined to be the maximum value of all lower
bounds (min_tsize) of each individual XferElement.

Call this function inside xfer_start(), after the individual elements have been
successfully initialized and before the transfer effectively starts.