device-src/device.c: further simplify device name validation
The 'errmsg' argument has become unnecessary: since we don't use a regex anymore
in validate_device_name(), the only possible error is an invalid device name.
As such, remove the argument and construct the error string in device_open()
instead.

while we're at it, make all error messages uniform in this function, and modify
the relevant tests accordingly.