Get rid of SIZEOF
The original purpose of this macro was to work around broken compilers (which
ones?) for which the return type of sizeof() was int.

Just get rid of it: newer code uses sizeof() anyway and nobody has complained
about this code breaking for them, so say goodbye to this 20th century artifact.