amanda-src/amandad.c: append to the buffer in case of read error
If read() failed, a string was written into buf but never written to the error
buffer. Fix that.

Take the opportunity to use a GString to hold the error buffer before it is
written to, since it simplifies the code (you can g_string_new(NULL) since GLib
will assume an empty string).