[allocapi] server-src/amcheck.c: replace strappend() usage, 3/3
In start_host(), strappend() was used to build the request to send. Use a
GString to build the request instead. This also allows to get rid of the
"number" local variable and push down the "req" local variable.

Also remove two unused local variables, which were assigned to but never used
for anything.